Added -RT option in weather to avoid the cancellation of an existing simulations folder

+++ weather.py
@@ -34,6 +34,8 @@ def read_arguments():
         default=999,
         help="Start date of the sampling period. Format: DD/MM/YYYY",
     )
+    parser.add_argument('-RT', '--run_type', default='new', help='Specify if the simulation is a new one or a restart.'
+                                                                 'Possible options are: new, restart')
     parser.add_argument(
         "-SY",
         "--sampled_years",
@@ -94,6 +96,7 @@ def read_arguments():
     )
     args = parser.parse_args()
     mode = args.mode
+    run_type = args.run_type
     start_date = args.start_date
     end_date = args.end_date
     sampled_years = args.sampled_years
@@ -113,6 +116,10 @@ def read_arguments():
     mode = mode.lower()
     if mode != "reanalysis" and mode != "forecast":
         print("ERROR. Wrong value for variable -M --mode")
+        sys.exit()
+    run_type = run_type.lower()
+    if run_type != 'new' and run_type != 'restart':
+        print('ERROR. Please provide a valid entry for -RT --run_type')
         sys.exit()
     if ERA5_on.lower() == "true":
         ERA5_on = True
@@ -221,6 +228,7 @@ def read_arguments():
         sys.exit()
     return (
         mode,
+        run_type,
         nsamples,
         time_start,
         time_stop,
@@ -1598,6 +1606,7 @@ def automatic_weather(analysis_start):
 
 (
     mode,
+    run_type,
     nsamples,
     time_start,
     time_stop,
@@ -1645,9 +1654,10 @@ try:
     os.mkdir(simulations)
 except FileExistsError:
     print("Folder " + simulations + " already exists")
-    for filename in os.listdir(simulations):
-        file_path = os.path.join(simulations, filename)
-        rmtree(file_path)
+    if run_type == 'new':
+        for filename in os.listdir(simulations):
+            file_path = os.path.join(simulations, filename)
+            rmtree(file_path)
 if os.path.exists("log.txt"):
     os.remove("log.txt")
 if os.path.exists("days_list.txt"):
